Feat: 육각형 데이터 조회 및 피드백 생성 로직 구현
프로젝트별 각 역량에 대한 평균 점수 계산 기능 구현
Description 매핑을 통한 역량 코드와 이름 연동
전체 사용자 평균과 비교하여 피드백 생성 로직 추가
코드 리팩토링 (메소드 책임 분리)

Resolves: S12P11A605-342

+++ collabit-server/src/main/java/com/collabit/project/domain/dto/SkillData.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Builder
 @Getter
 public class SkillData {
-    int score;
+    Double score;
     String name;
     String description;
 }
